+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add TestUtils to be able to use the PersonProvider in app tests
 - Improve caching of current user
 - Add new methods isCurrentUserAnEmployee/isCurrentUserAStudent/isCurrentUserAnAlumni